requirements.txt
pandas
openai
google-generativeai
python-dotenv
fastapi
uvicorn
yagmail
psycopg2-binary
